test for [DROOLS-5549]
