+++ drools-model/drools-model-compiler/src/test/java/org/drools/modelcompiler/FromTest.java
@@ -16,12 +16,7 @@
 
 package org.drools.modelcompiler;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 import org.assertj.core.api.Assertions;
 import org.drools.modelcompiler.FunctionsTest.Pojo;
@@ -625,5 +620,114 @@ public class FromTest extends BaseModelTest {
         ksession.fireAllRules();
 
         Assertions.assertThat(list).contains("Toystore1");
+    }
+
+    public static class Measurement {
+        private String id;
+        private String val;
+
+        public Measurement(String id, String val) {
+            this.id = id;
+            this.val = val;
+        }
+
+        public String getId() {
+            return id;
+        }
+
+        public void setId(String id) {
+            this.id = id;
+        }
+
+        public String getVal() {
+            return val;
+        }
+
+        public void setVal(String val) {
+            this.val = val;
+        }
+    }
+
+    @Test
+    public void testFromFunctionCall() {
+        // DROOLS-5548
+        String str =
+                "package com.sample;" +
+                        "global java.util.Set controlSet;\n" +
+                        "import " + Measurement.class.getCanonicalName() + ";\n" +
+                        "" +
+                        "declare A\n" +
+                        " x: String\n" +
+                        "end\n" +
+                        "" +
+                        "declare B\n" +
+                        " a: A\n" +
+                        "end\n" +
+                        "" +
+                        "function String dummyFunction(A b) {\n" +
+                        " return \"test\";\n" +
+                        "}\n" +
+                        "\n" +
+                        "rule \"insertB\"\n" +
+                        "when\n" +
+                        "then\n" +
+                        "drools.insert(new B(new A()));" +
+                        "end;" +
+                        "rule \"will execute per each Measurement having ID color\"\n" +
+                        "no-loop\n" +
+                        "when\n" +
+                        " Measurement( id == \"color\", $colorVal : val )\n" +
+                        " $b: B()\n" +
+                        " $val: String() from dummyFunction($b.a)\n" +
+                        "then\n" +
+                        " controlSet.add($colorVal);\n" +
+                        "end";
+
+        KieSession ksession = getKieSession( str );
+
+        HashSet<Object> hashSet = new HashSet<>();
+        ksession.setGlobal("controlSet", hashSet);
+
+        ksession.insert(new Measurement("color", "red"));
+
+        int ruleFired = ksession.fireAllRules();
+
+        assertEquals( 2, ruleFired );
+        assertEquals( "red", hashSet.iterator().next() );
+    }
+
+    @Test
+    public void testFromMap() {
+        // DROOLS-5549
+        String str =
+                "package com.sample;" +
+                        "global java.util.Set controlSet;\n" +
+                        "import " + Measurement.class.getCanonicalName() + ";\n" +
+                        "import " + Map.class.getCanonicalName() + ";\n" +
+                        "" +
+                        "function String dummyFunction(Map m) {\n" +
+                        " return \"test\";\n" +
+                        "}" +
+                        "\n" +
+                        "rule \"will execute per each Measurement having ID color\"\n" +
+                        "no-loop\n" +
+                        "when\n" +
+                        " Measurement( id == \"color\", $colorVal : val )\n" +
+                        " $val: String() from dummyFunction(Map.of($colorVal, \"something\"))\n" +
+                        "then\n" +
+                        " controlSet.add($colorVal);\n" +
+                        "end";
+
+        KieSession ksession = getKieSession( str );
+
+        HashSet<Object> hashSet = new HashSet<>();
+        ksession.setGlobal("controlSet", hashSet);
+
+        ksession.insert(new Measurement("color", "red"));
+
+        int ruleFired = ksession.fireAllRules();
+
+        assertEquals( 1, ruleFired );
+        assertEquals( "red", hashSet.iterator().next() );
     }
 }
