+++ drools-model/drools-model-compiler/src/test/java/org/drools/modelcompiler/FunctionsTest.java
@@ -259,78 +259,4 @@ public class FunctionsTest extends BaseModelTest {
         assertEquals( 2, results.size() );
         assertEquals( "whatever", results.iterator().next().getValue() );
     }
-
-    public static class Measurement {
-        private String id;
-        private String val;
-
-        public Measurement(String id, String val) {
-            this.id = id;
-            this.val = val;
-        }
-
-        public String getId() {
-            return id;
-        }
-
-        public void setId(String id) {
-            this.id = id;
-        }
-
-        public String getVal() {
-            return val;
-        }
-
-        public void setVal(String val) {
-            this.val = val;
-        }
-    }
-
-    @Test
-    public void testVariableAccessors() {
-        // DROOLS-5548
-        String str =
-                "package com.sample;" +
-                "global java.util.Set controlSet;\n" +
-                "import " + Measurement.class.getCanonicalName() + ";\n" +
-                "" +
-                "declare A\n" +
-                " x: String\n" +
-                "end\n" +
-                "" +
-                "declare B\n" +
-                " a: A\n" +
-                "end\n" +
-                "" +
-                "function String dummyFunction(A b) {\n" +
-                " return \"test\";\n" +
-                "}\n" +
-                "\n" +
-                "rule \"insertB\"\n" +
-                "when\n" +
-                "then\n" +
-                    "drools.insert(new B(new A()));" +
-                "end;" +
-                "rule \"will execute per each Measurement having ID color\"\n" +
-                "no-loop\n" +
-                "when\n" +
-                " Measurement( id == \"color\", $colorVal : val )\n" +
-                " $b: B()\n" +
-                " $val: String() from dummyFunction($b.a)\n" +
-                "then\n" +
-                " controlSet.add($colorVal);\n" +
-                "end";
-
-        KieSession ksession = getKieSession( str );
-
-        HashSet<Object> hashSet = new HashSet<>();
-        ksession.setGlobal("controlSet", hashSet);
-
-        ksession.insert(new Measurement("color", "red"));
-
-        int ruleFired = ksession.fireAllRules();
-
-        assertEquals( 2, ruleFired );
-        assertEquals( "red", hashSet.iterator().next() );
-    }
 }
